requirements.txt
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyparsing==3.2.0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
spotipy==2.24.0
uritemplate==4.1.1
urllib3==2.2.3
youtube-dl==2021.12.17
